opcua_client/MQTT/requirements.txt
asyncua==1.1.5
PyYAML==5.3.1
requests==2.26.0
paho-mqtt==2.1.0